requirements_web.txt
# Web应用依赖
flask>=2.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
slowapi>=0.1.0
python-multipart>=0.0.5

# 图像处理
opencv-python>=4.4
pillow>=8.0.0
numpy>=1.21.0

# 工具库
khandy>=0.1.8
onnxruntime>=1.20.0

# 配置和日志
pyyaml>=5.4.0
python-dotenv>=0.19.0

# 可选：数据库支持
# sqlalchemy>=1.4.0
# alembic>=1.7.0

# 可选：缓存支持
# redis>=4.0.0 